refactoring/improvements, next: making it so api_keys pulled off of firebase_db
